Moved the input_processor.py file to the utils module
